+++ packages/teleterm/src/ui/StatusBar/ShareFeedback/ShareFeedbackForm.tsx
@@ -1,100 +1,124 @@
 import React from 'react';
-import FieldInput from 'shared/components/FieldInput';
-import { requiredField } from 'shared/components/Validation/rules';
-import { FieldTextArea } from 'shared/components/FieldTextArea';
-import { Text } from 'design';
-import Toggle from 'teleport/components/Toggle';
+import { ButtonIcon, ButtonPrimary, Flex, Link, Text } from 'design';
+import Validation from 'shared/components/Validation';
+import { Close } from 'design/Icon';
+import { ShareFeedbackFormFields } from './ShareFeedbackFormFields';
+import { Attempt } from 'shared/hooks/useAsync';
+import * as Alerts from 'design/Alert';
+import { FEEDBACK_TOO_LONG_ERROR } from './useShareFeedback';
 import { ShareFeedbackFormValues } from './types';
 
-interface ShareFeedbackFormProps {
-  disabled: boolean;
+interface ShareFeedbackProps {
+  submitFeedbackAttempt: Attempt<Response>;
   formValues: ShareFeedbackFormValues;
 
+  onClose(): void;
+
   setFormValues(values: ShareFeedbackFormValues): void;
+
+  submitFeedback(): Promise<[Response, Error]>;
 }
 
-export function ShareFeedbackForm({
-  formValues,
-  setFormValues,
-  disabled,
-}: ShareFeedbackFormProps) {
-  function updateFormField<T extends keyof ShareFeedbackFormValues>(
-    field: T,
-    value: ShareFeedbackFormValues[T]
-  ) {
-    setFormValues({ ...formValues, [field]: value });
+export function ShareFeedbackForm(props: ShareFeedbackProps) {
+  const isSubmitInProgress =
+    props.submitFeedbackAttempt.status === 'processing';
+
+  return (
+    <Flex bg="primary.main" p={3} borderRadius={3} maxWidth="370px">
+      <Validation>
+        {({ validator }) => (
+          <Flex
+            flexDirection="column"
+            as="form"
+            onSubmit={e => {
+              e.preventDefault();
+              if (validator.validate()) {
+                props.submitFeedback();
+              }
+            }}
+          >
+            <Flex justifyContent="space-between" mb={2}>
+              <Text typography="h4" bold color="text.primary">
+                Provide your feedback
+              </Text>
+              <ButtonIcon
+                type="button"
+                onClick={props.onClose}
+                title="Close"
+                color="text.secondary"
+              >
+                <Close fontSize={5} />
+              </ButtonIcon>
+            </Flex>
+            <Link
+              href="https://github.com/gravitational/teleport/issues/new?assignees=&labels=bug&template=bug_report.md"
+              target="_blank"
+            >
+              Submit a Bug
+            </Link>
+            <Link href="https://goteleport.com/signup/" target="_blank">
+              Try Teleport Cloud
+            </Link>
+            {props.submitFeedbackAttempt.status === 'error' && (
+              <SubmissionError
+                submitFeedbackAttempt={props.submitFeedbackAttempt}
+              />
+            )}
+            {props.submitFeedbackAttempt.status === 'success' ? (
+              <Alerts.Success mt={3} mb={0}>
+                Your feedback has been submitted. Thank you!
+              </Alerts.Success>
+            ) : (
+              <>
+                <ShareFeedbackFormFields
+                  disabled={isSubmitInProgress}
+                  formValues={props.formValues}
+                  setFormValues={props.setFormValues}
+                />
+                <ButtonPrimary
+                  disabled={isSubmitInProgress}
+                  block
+                  type="submit"
+                  mt={4}
+                >
+                  Submit
+                </ButtonPrimary>
+              </>
+            )}
+          </Flex>
+        )}
+      </Validation>
+    </Flex>
+  );
+}
+
+function SubmissionError(props: { submitFeedbackAttempt: Attempt<Response> }) {
+  function getErrorText() {
+    if (props.submitFeedbackAttempt.statusText === FEEDBACK_TOO_LONG_ERROR) {
+      return (
+        <span>
+          That's a very long reason why. Please let us know more in{' '}
+          <Link
+            href="https://github.com/gravitational/teleport/discussions"
+            target="_blank"
+          >
+            our community
+          </Link>
+          .
+        </span>
+      );
+    }
+
+    return (
+      <span>
+        Unable to submit your feedback: {props.submitFeedbackAttempt.statusText}
+      </span>
+    );
   }
 
   return (
-    <>
-      <FieldInput
-        mt={3}
-        type="email"
-        label="Email Address"
-        autoFocus
-        readonly={disabled}
-        css={`
-          input {
-            font-size: 14px;
-          }
-        `}
-        rule={requiredField('Email is required')}
-        value={formValues.email}
-        onChange={e => updateFormField('email', e.target.value)}
-      />
-      <FieldInput
-        label="Company name (optional)"
-        autoFocus
-        readonly={disabled}
-        css={`
-          input {
-            font-size: 14px;
-          }
-        `}
-        value={formValues.company}
-        onChange={e => updateFormField('company', e.target.value)}
-      />
-      <FieldTextArea
-        label="Suggestions (Optional)"
-        textAreaCss={`
-                font-size: 14px;
-              `}
-        readOnly={disabled}
-        value={formValues.feedback}
-        onChange={e => updateFormField('feedback', e.target.value)}
-        placeholder="Type your suggestions here"
-      />
-      <Toggle
-        disabled={disabled}
-        isToggled={formValues.newsletterEnabled}
-        onToggle={() => {
-          updateFormField('newsletterEnabled', !formValues.newsletterEnabled);
-        }}
-      >
-        <Text ml={2} color="text.primary">
-          Sign me up for the newsletter
-        </Text>
-      </Toggle>
-      <Toggle
-        disabled={disabled}
-        isToggled={formValues.salesContactEnabled}
-        onToggle={() => {
-          updateFormField(
-            'salesContactEnabled',
-            !formValues.salesContactEnabled
-          );
-        }}
-      >
-        <Text
-          ml={2}
-          color="text.primary"
-          css={`
-            line-height: 18px;
-          `}
-        >
-          I would like a demo of Teleport&nbsp;Enterprise features
-        </Text>
-      </Toggle>
-    </>
+    <Alerts.Danger mt={3} mb={0}>
+      {getErrorText()}
+    </Alerts.Danger>
   );
 }
