+++ packages/teleterm/src/ui/StatusBar/ShareFeedback/index.ts
@@ -1,1 +1,1 @@
-export * from './ShareFeedbackContainer';
+export * from './ShareFeedback';
